requirements.txt
django== 5.1
djangorestframework
setuptools
djangorestframework-simplejwt
django-auditlog
mysqlclient
drf-spectacular
pandas
pika
celery
django-filter